♻️ REFACTOR: 달력, 캔버스 폴더 경로를 common폴터 안으로 이동
