+++ src/components/common/calendar/YearSelectModal.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useCallback } from "react";
 import styled from "styled-components";
-import { Modal } from "../common/modal/ReactModal";
+import { Modal } from "../modal/ReactModal";
 
 const YearSelectModal = ({ children, onYearController }) => {
   const years = useCallback(() => {
